refactor(minesweeper): Builder 를 dsl 패키지로 이동
