+++ src/main/kotlin/tdd/minesweeper/domain/dsl/Builder.kt
@@ -1,4 +1,4 @@
-package tdd.minesweeper.domain
+package tdd.minesweeper.domain.dsl
 
 interface Builder<T> {
     fun build(): T
